+++ polkadot/roadmap/implementers-guide/src/types/overseer-protocol.md
@@ -901,22 +901,6 @@ const APPROVAL_EXECUTION_TIMEOUT: Duration = 6 seconds;
 /// or `Ok(ValidationResult::Invalid)`.
 #[derive(Debug)]
 pub enum CandidateValidationMessage {
-    /// Validate a candidate with provided parameters using relay-chain state.
-    ///
-    /// This will implicitly attempt to gather the `PersistedValidationData` and `ValidationCode`
-    /// from the runtime API of the chain, based on the `relay_parent`
-    /// of the `CandidateDescriptor`.
-    ///
-    /// This will also perform checking of validation outputs against the acceptance criteria.
-    ///
-    /// If there is no state available which can provide this data or the core for
-    /// the para is not free at the relay-parent, an error is returned.
-    ValidateFromChainState(
-        CandidateDescriptor,
-        Arc<PoV>,
-        Duration, // Execution timeout.
-        oneshot::Sender<Result<ValidationResult, ValidationFailed>>,
-    ),
     /// Validate a candidate with provided, exhaustive parameters for validation.
     ///
     /// Explicitly provide the `PersistedValidationData` and `ValidationCode` so this can do full
